pyproject.toml
[project]
name = "issue-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.0",
    "dotenv>=0.9.9",
    "fast-agent-mcp>=0.2.58",
    "fastapi>=0.120.2",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=0.0.3",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.1",
    "openai>=2.6.1",
    "pygithub>=2.8.1",
    "sentence-transformers>=5.1.2",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
]
